+++ trunk/xml/src/jlibs/xml/sax/sniff/model/expr/nodeset/list/NodeList.java
@@ -58,7 +58,7 @@ public abstract class NodeList extends ValidatedExpression{
                     }
 
                     @Override
-                    public void contextEnded(Context context, int order){
+                    public void contextEnded(Context context, long order){
                         for(Evaluation eval: evaluationStack){
                             NodeListEvaluation evaluation = (NodeListEvaluation)eval;
                             if(!evaluation.finished && !evaluation.resultPrepared){
